+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - [client] App name length validation
 
+### Changed
+- [server] Specific CPU and Memory limits for both deploy and release pods
+
 ## [0.5.0] - 2017-08-15
 ### Changed
 - Use gRPC instead of go-swagger